i18n(cape-ban): localize ban-expiry timestamp formatting
Extract the inline ban-timestamp formatter into a reusable
humanizeTimestamps() in time-utils, driven by new time.* keys across
all 10 locales instead of hardcoded English.
