+++ src/i18n/translations/de.json
@@ -2013,5 +2013,11 @@
   "crash_analysis.summary.loader_update": "Aktualisiere {{mod}} auf {{target}}.",
   "crash_analysis.summary.install": "Installiere {{mod}} {{target}}, eine Mod braucht sie.",
   "crash_analysis.summary.disable_mod": "Deaktiviere \"{{mod}}\", das ist die fehlerhafte Mod.",
-  "crash_analysis.summary.multi": "{{count}} Fixes können das lösen. Wende sie unten an."
+  "crash_analysis.summary.multi": "{{count}} Fixes können das lösen. Wende sie unten an.",
+  "time.permanent": "dauerhaft",
+  "time.until_expired": "bis {{date}} (abgelaufen)",
+  "time.until_remaining": "bis {{date}} (noch {{remaining}})",
+  "time.remaining_days_hours": "{{days}}T {{hours}}Std",
+  "time.remaining_hours_minutes": "{{hours}}Std {{minutes}}Min",
+  "time.remaining_minutes": "{{minutes}}Min"
 }
