+++ src/i18n/translations/en.json
@@ -2031,5 +2031,11 @@
   "crash_analysis.summary.loader_update": "Update {{mod}} to {{target}}.",
   "crash_analysis.summary.install": "Install {{mod}} {{target}}, a mod needs it.",
   "crash_analysis.summary.disable_mod": "Disable \"{{mod}}\", it's the failing mod.",
-  "crash_analysis.summary.multi": "{{count}} fixes can resolve this. Apply them below."
+  "crash_analysis.summary.multi": "{{count}} fixes can resolve this. Apply them below.",
+  "time.permanent": "permanently",
+  "time.until_expired": "until {{date}} (expired)",
+  "time.until_remaining": "until {{date}} ({{remaining}} left)",
+  "time.remaining_days_hours": "{{days}}d {{hours}}h",
+  "time.remaining_hours_minutes": "{{hours}}h {{minutes}}m",
+  "time.remaining_minutes": "{{minutes}}m"
 }
