+++ src/i18n/translations/pl.json
@@ -1848,5 +1848,11 @@
   "crash_analysis.summary.loader_update": "Zaktualizuj {{mod}} do {{target}}.",
   "crash_analysis.summary.install": "Zainstaluj {{mod}} {{target}}, mod tego wymaga.",
   "crash_analysis.summary.disable_mod": "Wyłącz \"{{mod}}\", to jest wadliwy mod.",
-  "crash_analysis.summary.multi": "{{count}} poprawek może to rozwiązać. Zastosuj je poniżej."
+  "crash_analysis.summary.multi": "{{count}} poprawek może to rozwiązać. Zastosuj je poniżej.",
+  "time.permanent": "na stałe",
+  "time.until_expired": "do {{date}} (wygasło)",
+  "time.until_remaining": "do {{date}} (pozostało {{remaining}})",
+  "time.remaining_days_hours": "{{days}}d {{hours}}godz",
+  "time.remaining_hours_minutes": "{{hours}}godz {{minutes}}min",
+  "time.remaining_minutes": "{{minutes}}min"
 }
