+++ src/i18n/translations/tr.json
@@ -1848,5 +1848,11 @@
   "crash_analysis.summary.loader_update": "{{mod}} modunu {{target}} sürümüne güncelle.",
   "crash_analysis.summary.install": "{{mod}} {{target}} kur, bir mod buna ihtiyaç duyuyor.",
   "crash_analysis.summary.disable_mod": "\"{{mod}}\" modunu devre dışı bırak, sorun çıkaran mod bu.",
-  "crash_analysis.summary.multi": "{{count}} çözüm bunu giderebilir. Aşağıdan uygula."
+  "crash_analysis.summary.multi": "{{count}} çözüm bunu giderebilir. Aşağıdan uygula.",
+  "time.permanent": "kalıcı olarak",
+  "time.until_expired": "{{date}} tarihine kadar (süresi doldu)",
+  "time.until_remaining": "{{date}} tarihine kadar ({{remaining}} kaldı)",
+  "time.remaining_days_hours": "{{days}}g {{hours}}sa",
+  "time.remaining_hours_minutes": "{{hours}}sa {{minutes}}dk",
+  "time.remaining_minutes": "{{minutes}}dk"
 }
